+++ src/compose.ts
@@ -9,7 +9,7 @@ import { parse as parseQuery } from 'fast-querystring'
 // @ts-expect-error
 import decodeURIComponent from 'fast-decode-uri-component'
 
-import { signCookie } from './utils'
+import { getCookieValidator, lifeCycleToFn, signCookie } from './utils'
 import { ParseError, error } from './error'
 
 import {
@@ -32,6 +32,7 @@ import { parseCookie, type CookieOptions } from './cookies'
 import type {
 	ComposedHandler,
 	Handler,
+	HookContainer,
 	LifeCycleStore,
 	SchemaValidator,
 	TraceEvent
@@ -308,7 +309,9 @@ const getUnionedType = (validator: TypeCheck<any> | undefined) => {
 
 const matchFnReturn = /(?:return|=>) \S+\(/g
 
-export const isAsync = (fn: Function) => {
+export const isAsync = (v: Function | HookContainer) => {
+	const fn = 'fn' in v ? v.fn : v
+
 	if (fn.constructor.name === 'AsyncFunction') return true
 
 	const literal = fn.toString()
@@ -344,7 +347,12 @@ export const composeHandler = ({
 	}
 }): ComposedHandler => {
 	const isHandleFn = typeof handler === 'function'
-	if (!isHandleFn) handler = mapCompactResponse(handler)
+
+	if (!isHandleFn)
+		handler = mapResponse(handler, {
+			// @ts-expect-error private property
+			headers: app.setHeaders ?? {}
+		})
 
 	const hasErrorHandler =
 		(app.config.forceErrorEncapsulation &&
@@ -399,8 +407,20 @@ export const composeHandler = ({
 	const hasHeaders = inference.headers || validator.headers
 	const hasCookie = inference.cookie || !!validator.cookie
 
+	const cookieValidator = hasCookie
+		? getCookieValidator({
+				validator: validator.cookie as any,
+				defaultConfig: app.config.cookie,
+				dynamic: !!app.config.aot,
+				// @ts-expect-error
+				config: validator.cookie?.config ?? {},
+				// @ts-expect-error
+				models: app.definitions.type
+		  })
+		: undefined
+
 	// @ts-ignore private property
-	const cookieMeta = validator?.cookie?.schema as {
+	const cookieMeta = cookieValidator?.config as {
 		secrets?: string | string[]
 		sign: string[] | true
 		properties: { [x: string]: Object }
@@ -563,7 +583,7 @@ export const composeHandler = ({
 		inference.set ||
 		hasTraceSet ||
 		hasHeaders ||
-		hasDefaultHeaders
+		(isHandleFn && hasDefaultHeaders)
 
 	if (hasTrace) fnLiteral += '\nconst id = c.$$requestId\n'
 
@@ -612,10 +632,10 @@ export const composeHandler = ({
 					case 'json':
 					case 'application/json':
 						if (hasErrorHandler)
-							fnLiteral += `const body = await c.request.text()
+							fnLiteral += `const tempBody = await c.request.text()
 							
 							try {
-								c.body = JSON.parse(body)
+								c.body = JSON.parse(tempBody)
 							} catch {
 								throw new ParseError('Failed to parse body as found: ' + (typeof body === "string" ? "'" + body + "'" : body), body)
 							}`
@@ -703,7 +723,7 @@ export const composeHandler = ({
 
 					for (let i = 0; i < hooks.parse.length; i++) {
 						const endUnit = report('parse.unit', {
-							name: hooks.parse[i].name
+							name: hooks.parse[i].fn.name
 						})
 
 						const name = `bo${i}`
@@ -730,10 +750,10 @@ export const composeHandler = ({
 						${
 							hasErrorHandler
 								? `
-						const body = await c.request.text()
+						const tempBody = await c.request.text()
 						
 						try {
-							c.body = JSON.parse(body)
+							c.body = JSON.parse(tempBody)
 						} catch {
 							throw new ParseError('Failed to parse body as found: ' + (typeof body === "string" ? "'" + body + "'" : body), body)
 						}
@@ -785,21 +805,23 @@ export const composeHandler = ({
 			unit: hooks.transform.length
 		})
 
+		if (hooks.transform.length) fnLiteral += '\nlet transformed\n'
+
 		for (let i = 0; i < hooks.transform.length; i++) {
 			const transform = hooks.transform[i]
 
 			const endUnit = report('transform.unit', {
-				name: transform.name
+				name: transform.fn.name
 			})
 
-			if (transform.$elysia === 'derive')
-				fnLiteral += isAsync(transform)
-					? `Object.assign(c, await transform[${i}](c));`
-					: `Object.assign(c, transform[${i}](c));`
+			fnLiteral += isAsync(transform)
+				? `transformed = await transform[${i}](c)\n`
+				: `transformed = transform[${i}](c)\n`
+
+			fnLiteral += `if(transformed?.[ELYSIA_RESPONSE])
+				throw transformed
 			else
-				fnLiteral += isAsync(transform)
-					? `await transform[${i}](c);`
-					: `transform[${i}](c);`
+				Object.assign(c, transformed)\n`
 
 			endUnit()
 		}
@@ -922,17 +944,17 @@ export const composeHandler = ({
 		}
 
 		// @ts-ignore
-		if (isNotEmpty(validator.cookie?.schema.properties ?? {})) {
+		if (isNotEmpty(cookieValidator?.schema.properties ?? {})) {
 			fnLiteral += `const cookieValue = {}
     			for(const [key, value] of Object.entries(c.cookie))
     				cookieValue[key] = value.value\n`
 
 			// @ts-ignore
-			if (hasProperty('default', validator.cookie.schema))
+			if (hasProperty('default', cookieValidator.schema))
 				for (const [key, value] of Object.entries(
 					Value.Default(
 						// @ts-ignore
-						validator.cookie.schema,
+						cookieValidator.schema,
 						{}
 					) as Object
 				)) {
@@ -958,20 +980,33 @@ export const composeHandler = ({
 			unit: hooks.beforeHandle.length
 		})
 
+		let hasResolve = false
+
 		for (let i = 0; i < hooks.beforeHandle.length; i++) {
 			const beforeHandle = hooks.beforeHandle[i]
 
 			const endUnit = report('beforeHandle.unit', {
-				name: beforeHandle.name
+				name: beforeHandle.fn.name
 			})
 
-			const returning = hasReturn(beforeHandle.toString())
+			const returning = hasReturn(beforeHandle.fn.toString())
 
-			// @ts-ignore
-			if (beforeHandle.$elysia === 'resolve') {
+			const isResolver = beforeHandle.subType === 'resolve'
+
+			if (isResolver) {
+				if (!hasResolve) {
+					hasResolve = true
+					fnLiteral += '\nlet resolved\n'
+				}
+
 				fnLiteral += isAsync(beforeHandle)
-					? `Object.assign(c, await beforeHandle[${i}](c));`
-					: `Object.assign(c, beforeHandle[${i}](c));`
+					? `resolved = await beforeHandle[${i}](c);\n`
+					: `resolved = beforeHandle[${i}](c);\n`
+
+				fnLiteral += `if(resolved[ELYSIA_RESPONSE])
+						throw resolved
+					else
+						Object.assign(c, resolved)\n`
 			} else if (!returning) {
 				fnLiteral += isAsync(beforeHandle)
 					? `await beforeHandle[${i}](c);\n`
@@ -979,6 +1014,7 @@ export const composeHandler = ({
 
 				endUnit()
 			} else {
+				fnLiteral += `Object.assign(c, be);`
 				fnLiteral += isAsync(beforeHandle)
 					? `be = await beforeHandle[${i}](c);\n`
 					: `be = beforeHandle[${i}](c);\n`
@@ -997,22 +1033,22 @@ export const composeHandler = ({
 					})()
 
 					for (let i = 0; i < hooks.afterHandle.length; i++) {
-						const returning = hasReturn(
-							hooks.afterHandle[i].toString()
-						)
+						const hook = hooks.afterHandle[i]
+
+						const returning = hasReturn(hook.fn.toString())
 
 						const endUnit = report('afterHandle.unit', {
-							name: hooks.afterHandle[i].name
+							name: hook.fn.name
 						})
 
 						fnLiteral += `c.response = be\n`
 
 						if (!returning) {
-							fnLiteral += isAsync(hooks.afterHandle[i])
+							fnLiteral += isAsync(hook.fn)
 								? `await afterHandle[${i}](c, be)\n`
 								: `afterHandle[${i}](c, be)\n`
 						} else {
-							fnLiteral += isAsync(hooks.afterHandle[i])
+							fnLiteral += isAsync(hook.fn)
 								? `af = await afterHandle[${i}](c)\n`
 								: `af = afterHandle[${i}](c)\n`
 
@@ -1070,20 +1106,22 @@ export const composeHandler = ({
 		})
 
 		for (let i = 0; i < hooks.afterHandle.length; i++) {
-			const returning = hasReturn(hooks.afterHandle[i].toString())
+			const hook = hooks.afterHandle[i]
+
+			const returning = hasReturn(hook.fn.toString())
 
 			const endUnit = report('afterHandle.unit', {
-				name: hooks.afterHandle[i].name
+				name: hook.fn.name
 			})
 
 			if (!returning) {
-				fnLiteral += isAsync(hooks.afterHandle[i])
+				fnLiteral += isAsync(hook.fn)
 					? `await afterHandle[${i}](c)\n`
 					: `afterHandle[${i}](c)\n`
 
 				endUnit()
 			} else {
-				fnLiteral += isAsync(hooks.afterHandle[i])
+				fnLiteral += isAsync(hook.fn)
 					? `af = await afterHandle[${i}](c)\n`
 					: `af = afterHandle[${i}](c)\n`
 
@@ -1153,9 +1191,8 @@ export const composeHandler = ({
 			fnLiteral += encodeCookie
 
 			if (handler instanceof Response) {
-				fnLiteral +=
-					inference.set || hasDefaultHeaders
-						? `if(
+				fnLiteral += inference.set
+					? `if(
 					isNotEmpty(c.set.headers) ||
 					c.set.status !== 200 ||
 					c.set.redirect ||
@@ -1164,7 +1201,7 @@ export const composeHandler = ({
 					return mapResponse(${handle}.clone(), c.set, c.request)
 				else
 					return ${handle}.clone()`
-						: `return ${handle}.clone()`
+					: `return ${handle}.clone()`
 
 				fnLiteral += '\n'
 			} else if (hasSet)
@@ -1202,9 +1239,8 @@ export const composeHandler = ({
 			report('afterHandle')()
 
 			if (handler instanceof Response) {
-				fnLiteral +=
-					inference.set || hasDefaultHeaders
-						? `if(
+				fnLiteral += inference.set
+					? `if(
 					isNotEmpty(c.set.headers) ||
 					c.set.status !== 200 ||
 					c.set.redirect ||
@@ -1213,7 +1249,7 @@ export const composeHandler = ({
 					return mapResponse(${handle}.clone(), c.set, c.request)
 				else
 					return ${handle}.clone()`
-						: `return ${handle}.clone()`
+					: `return ${handle}.clone()`
 
 				fnLiteral += '\n'
 			} else if (hasSet)
@@ -1242,7 +1278,7 @@ export const composeHandler = ({
 			for (let i = 0; i < hooks.error.length; i++) {
 				const name = `er${i}`
 				const endUnit = report('error.unit', {
-					name: hooks.error[i].name
+					name: hooks.error[i].fn.name
 				})
 
 				fnLiteral += `\nlet ${name} = handleErrors[${i}](c)\n`
@@ -1347,7 +1383,7 @@ export const composeHandler = ({
 
 	return createHandler({
 		handler,
-		hooks,
+		hooks: lifeCycleToFn(hooks),
 		validator,
 		// @ts-expect-error
 		handleError: app.handleError,
@@ -1475,7 +1511,11 @@ export const composeGeneralHandler = (
 
 	const notFound = new NotFoundError()
 
-	${app.event.request.length ? `const onRequest = app.event.request` : ''}
+	${
+		app.event.request.length
+			? `const onRequest = app.event.request.map(x => x.fn)`
+			: ''
+	}
 	${router.static.http.variables}
 	${
 		app.event.error.length
@@ -1530,12 +1570,12 @@ export const composeGeneralHandler = (
 		fnLiteral += `\n try {\n`
 
 		for (let i = 0; i < app.event.request.length; i++) {
-			const fn = app.event.request[i]
-			const withReturn = hasReturn(fn.toString())
-			const maybeAsync = isAsync(fn)
+			const hook = app.event.request[i]
+			const withReturn = hasReturn(hook.fn.toString())
+			const maybeAsync = isAsync(hook)
 
 			const endUnit = report('request.unit', {
-				name: app.event.request[i].name
+				name: app.event.request[i].fn.name
 			})
 
 			if (withReturn) {
@@ -1665,11 +1705,14 @@ export const composeErrorHandler = (
 	app: Elysia<any, any, any, any, any, any, any, any>
 ) => {
 	let fnLiteral = `const {
-		app: { event: { error: onError, onResponse: res } },
+		app: { event: { error: onErrorContainer, onResponse: resContainer } },
 		mapResponse,
 		ERROR_CODE,
 		ELYSIA_RESPONSE
 	} = inject
+
+	const onError = onErrorContainer.map(x => x.fn)
+	const res = resContainer.map(x => x.fn)
 
 	return ${
 		app.event.error.find(isAsync) ? 'async' : ''
@@ -1695,7 +1738,7 @@ export const composeErrorHandler = (
 
 		fnLiteral += '\nif(skipGlobal !== true) {\n'
 
-		if (hasReturn(handler.toString()))
+		if (hasReturn(handler.fn.toString()))
 			fnLiteral += `r = ${response}; if(r !== undefined) {
 				if(r instanceof Response) return r