+++ src/sucrose.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable no-constant-condition */
-import type { Handler, LifeCycleStore, TraceHandler } from './types'
+import type { Handler, HookContainer, LifeCycleStore, TraceHandler } from './types'
 
 export namespace Sucrose {
 	export interface Inference {
@@ -324,6 +324,13 @@ export const findAlias = (type: string, body: string, depth = 0) => {
 	return aliases
 }
 
+const accessor = <T extends string, P extends string>(parent: T, prop: P) =>
+	[
+		parent + '.' + prop,
+		parent + '["' + prop + '"]',
+		parent + "['" + prop + "']"
+	] as const
+
 export const extractMainParameter = (parameter: string) => {
 	if (!parameter) return
 
@@ -357,6 +364,8 @@ export const inferBodyReference = (
 		code.includes(alias + "['" + type + "']")
 
 	for (let alias of aliases) {
+		if (!alias) continue
+
 		if (alias.charCodeAt(0) === 123) {
 			alias = retrieveRootParamters(alias)
 
@@ -391,7 +400,17 @@ export const inferBodyReference = (
 
 		if (!inference.query && access('query', alias)) inference.query = true
 
-		if (inference.query)
+		if (
+			code.includes('return ' + alias) ||
+			accessor('return ' + alias, 'query').some((key) =>
+				code.includes(key)
+			)
+		) {
+			inference.query = true
+			inference.unknownQueries = true
+		}
+
+		if (inference.query && !inference.unknownQueries)
 			while (true) {
 				let keyword = alias + '.'
 				if (code.includes(keyword + 'query')) keyword = alias + '.query'
@@ -664,7 +683,11 @@ export const sucrose = (
 	if (lifeCycle.request?.length) events.push(...lifeCycle.request)
 	if (lifeCycle.onResponse?.length) events.push(...lifeCycle.onResponse)
 
-	for (const event of events) {
+	for (const e of events) {
+		if(!e) continue
+		
+		const event = 'fn' in e ? e.fn : e
+
 		const [parameter, body] = separateFunction(event.toString())
 
 		const rootParameters = findParameterReference(parameter, inference)
@@ -675,6 +698,19 @@ export const sucrose = (
 			aliases.splice(0, -1, mainParameter)
 
 			inferBodyReference(body, aliases, inference)
+		}
+
+		const context = rootParameters || mainParameter
+		if (
+			context &&
+			['', 'return '].some((type) =>
+				accessor(type + context, 'query').some((key) =>
+					body.includes(key)
+				)
+			)
+		) {
+			inference.query = true
+			inference.unknownQueries = true
 		}
 
 		if (inference.query) {