+++ CHANGELOG.md
@@ -1,4 +1,63 @@
-# 1.0.0 - 8 Mar 2024
+# 1.0.7 - 20 Mar 2024
+Feature:
+- add Elysia.propagate to propagate hook type from 'local' to 'scoped'
+
+Improvement:
+- remove function.$elysia
+- remove function extension
+
+Bug fix:
+- duplicate macro call
+- [#548](https://github.com/elysiajs/elysia/issues/548) additional case for "accessing all query params using property name (ctx.query) doesn't work anymore"
+- [#599](https://github.com/elysiajs/elysia/issues/559) plugin with scoped settings not functioning correctly
+
+# 1.0.6 - 20 Mar 2024
+Bug fix:
+- inline function doesn't propagate correctly on type level
+
+# 1.0.5 - 19 Mar 2024
+Improvement:
+- using regex for date pattern matching before using new Date validation
+- using tsc to emit declaration file instead of tsup
+- add `mapResponse` to MacroManager
+
+Bug fix:
+- Ephemeral and Volatile type isn't recognize by MacroManager
+- inline guard cookie doesn't apply to local instance
+
+# 1.0.4 - 18 Mar 2024
+Improvement:
+- resolve, derive soundness
+
+# 1.0.3 - 18 Mar 2024
+Improvement:
+- Reduce instruction for static resource
+
+Bug fix:
+- Fix returning mulitple status code using `error` doesn't accept the response
+
+# 1.0.2 - 18 Mar 2024
+Feature:
+- add `scoped` support for `derive` and `resolve`
+
+Improvement:
+- Type soundness
+- type inference performance improvement
+
+# 1.0.1 - 18 Mar 2024
+Improvement:
+- `mapHandler` now check passthrough once instead of twice
+- exclude return type of`ELYSIA_RESPONSE` type from `derive` and `resolve`
+- throw error if `error` is return in `derive` and `resolve`
+- handle `return error` on `transform`
+- [#502](https://github.com/elysiajs/elysia/pull/502) merge response schema from parent scope
+
+Bug fix:
+- explicit `type: 'json'` with body schema throw unexpected `body.Check` is not a function
+- [#549](https://github.com/elysiajs/elysia/pull/549) await the .modules of nested Elysia instances
+- [#548](https://github.com/elysiajs/elysia/issues/548) Accessing all query params using property name (ctx.query) doesn't work anymore
+
+# 1.0.0 - 16 Mar 2024
 Improvement:
 - fine-grained reactive cookie
 - using single source of truth for cookie